+++ hcaptcha_challenger/constant.py
@@ -1,0 +1,35 @@
+# -*- coding: utf-8 -*-
+# Time       : 2024/4/19 2:29
+# Author     : QIN2DIM
+# GitHub     : https://github.com/QIN2DIM
+# Description:
+BAD_CODE = {
+    "а": "a",
+    "е": "e",
+    "e": "e",
+    "i": "i",
+    "і": "i",
+    "ο": "o",
+    "с": "c",
+    "ԁ": "d",
+    "ѕ": "s",
+    "һ": "h",
+    "у": "y",
+    "р": "p",
+    "ϳ": "j",
+    "х": "x",
+    "\u0405": "S",
+    "\u0042": "B",
+    "\u0052": "R",
+    "\u0049": "I",
+    "\u0043": "C",
+    "\u004b": "K",
+    "\u039a": "K",
+    "\u0053": "S",
+    "\u0421": "C",
+    "\u006c": "l",
+    "\u0399": "I",
+    "\u0392": "B",
+    "ー": "一",
+    "土": "士",
+}
