+++ hcaptcha_challenger/tools/__init__.py
@@ -3,3 +3,29 @@
 # Author     : QIN2DIM
 # GitHub     : https://github.com/QIN2DIM
 # Description:
+from .cv_toolkit import (
+    annotate_objects,
+    find_unique_object,
+    find_similar_objects,
+    find_unique_color,
+)
+from .image_downloader import download_challenge_images
+from .image_label_binary import rank_models, match_datalake
+from .match_model import match_model
+from .prompt_handler import handle
+from .zero_shot_image_classifier import ZeroShotImageClassifier, register_pipline, invoke_clip_tool
+
+__all__ = [
+    "download_challenge_images",
+    "rank_models",
+    "match_datalake",
+    "annotate_objects",
+    "find_unique_object",
+    "find_similar_objects",
+    "handle",
+    "ZeroShotImageClassifier",
+    "register_pipline",
+    "match_model",
+    "find_unique_color",
+    "invoke_clip_tool",
+]
